backend/requirements.txt
fastapi==0.115.2
uvicorn[standard]==0.30.6
SQLAlchemy==2.0.34
pydantic==2.9.2
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0
python-multipart==0.0.9
psycopg2-binary==2.9.9
alembic==1.13.1
python-dotenv==1.0.1